refactor(ButtonGroup): Extract component for future reuse
